Add JSON view
You can now toggle between a tree and a JSON view. The JSON view allows
you to inspect and copy the AST as JSON.

In the future, you will be able to paste and edit an existing AST through the
JSON view.

This also improves performance when working with large scripts, by
properly using CoreMirror.