Disable minification for flow parser; do not merge chunks
When applying uglifyjs to the flow parser, it produces octal escape
sequences which are not allowed in strict mode. Chrome seems to tolerate
this, but not Firefox, Safari or IE. To make the flow parser (and
anything that uses the flow parser, such as jscodeshift) work in these
browsers, we selectively disable minification for that file, using the
uglify loader instead of the plugin.

Since the minification is now applied to individual files instead of
bundles, the resulting files might be a bit larger, but hopefully that
doesn't have much impact after gzip.

---

This also disables merging of chunks if they contain the same module. I
believe that in reality, most people will use one or two parsers. With
the merging enabled, parsers that are built on top (e.g. acorn, flow,
esprima) have all been merged into one file, resulting in one huge file
that needed to be downloaded, even if only one of those parser was used.