Restructure menu; add more options
In order to accommodate a way to get URLs for sharing and to reset
content to the built-in code examples, "Save" and "Fork" are replaced by
a single "Snippet" menu.

The menu contains "Save" and "Fork" and two new buttons:

- "New": resets the state to the built-in examples and clears the
URL
- "Share...": opens a dialog that has copyable URLs for the current and
latest revisions

The "Snippet" toolbar entry also features a tiny save or fork button,
depending on which action is available (prioritizing save). So it's
still possible to "on-click" save a snippet.