Serve old Parse data from express server
Now that Parse shut down, the data is served from our own server
instead. The amount of data is so little (less than 30MB), that we can
simply load it into memory. Cache control will be handled by nginx.

The files containing the data are not committed here for privacy
reasons. Instead they are loaded via environment variables on the
production server.

The server can still be run locally without these files because the
corresponding code is not loaded if the environment variables are not
present.