Use flexbox for Splitpane implementation (fixes #498)
The combination of flexbox and `position: relative/absolute` caused
issues in Safari.
I always intended to move the Splitpane implementation to use flexbox
but hit a roadblock when it came to limiting the editors to the
available height. Turns out the solution is to use flexbox all the way
down and to set `min-width/height` in a couple of places to make sure
that the elements can grow towards the available space.